move server_options to app_config to fix wrong import direction
(components -> bases should be the other way around)
