+++ projects/background_jobs/pyproject.toml
@@ -23,6 +23,7 @@ packages = [
   { include = "renku_data_services/message_queue", from = "../../components" },
   { include = "renku_data_services/db_config", from = "../../components" },
   { include = "renku_data_services/k8s", from = "../../components" },
+  { include = "renku_data_services/k8s_watcher", from = "../../components" },
   { include = "renku_data_services/crc", from = "../../components" },
   { include = "renku_data_services/project", from = "../../components" },
   { include = "renku_data_services/authz", from = "../../components" },
