+++ test/bases/renku_data_services/data_api/test_parsing_old_server_options.py
@@ -2,12 +2,12 @@ from io import StringIO
 
 from yaml import safe_load
 
-from renku_data_services.crc import models
-from renku_data_services.data_api.server_options import (
+from renku_data_services.app_config.server_options import (
     ServerOptions,
     ServerOptionsDefaults,
     generate_default_resource_pool,
 )
+from renku_data_services.crc import models
 
 server_options_yaml = """
 defaultUrl:
