+++ test/utils.py
@@ -18,16 +18,16 @@ from yaml import safe_load
 import renku_data_services.base_models as base_models
 from renku_data_services import errors
 from renku_data_services.app_config.config import BuildsConfig, Config, SentryConfig, TrustedProxiesConfig
+from renku_data_services.app_config.server_options import (
+    ServerOptions,
+    ServerOptionsDefaults,
+    generate_default_resource_pool,
+)
 from renku_data_services.authn.dummy import DummyAuthenticator, DummyUserStore
 from renku_data_services.authz.authz import Authz
 from renku_data_services.authz.config import AuthzConfig
 from renku_data_services.crc import models as rp_models
 from renku_data_services.crc.db import ResourcePoolRepository
-from renku_data_services.data_api.server_options import (
-    ServerOptions,
-    ServerOptionsDefaults,
-    generate_default_resource_pool,
-)
 from renku_data_services.db_config.config import DBConfig
 from renku_data_services.git.gitlab import DummyGitlabAPI
 from renku_data_services.k8s.clients import DummyCoreClient, DummySchedulingClient
